Eslint for vue

+++ src/views/Users.vue
@@ -1,39 +1,59 @@
 <template>
   <div class="home">
-    <Nav></Nav>
+    <Nav />
     <div class="wrapper">
       <h2>Users</h2>
       <hr>
-      <a class="waves-effect waves-light btn-large" v-on:click="openUserInterface()">Add User</a>
+      <a
+        class="waves-effect waves-light btn-large"
+        @click="openUserInterface()"
+      >Add User</a>
       <hr>
-      
+
       <table>
         <thead>
           <tr>
-              <th>Name</th>
-              <th></th>
+            <th>Name</th>
+            <th />
           </tr>
         </thead>
 
         <tbody>
-          <tr v-for="user in allUsers" :key="user.id" :class="{editing: user == editedUser}" v-cloak>
+          <tr
+            v-for="user in allUsers"
+            v-cloak
+            :key="user.id"
+            :class="{editing: user == editedUser}"
+          >
             <td>
               <div class="view">
-                {{user.username}}
+                {{ user.username }}
               </div>
               <div class="edit">
-                <input type="text" v-model="user.username"/>
+                <input
+                  v-model="user.username"
+                  type="text"
+                >
               </div>
             </td>
             <td>
-                <div class="view">
-                  <a class="theB" v-on:click="editUser(user)">Edit</a>
-                </div>
-                
-                <div class="edit">
-                  <a class="theB" v-on:click="saveUser(user)">Save</a>
-                </div>
-                <a class="theB delete" v-on:click="deleteUser(user.id)">Delete</a>
+              <div class="view">
+                <a
+                  class="theB"
+                  @click="editUser(user)"
+                >Edit</a>
+              </div>
+
+              <div class="edit">
+                <a
+                  class="theB"
+                  @click="saveUser(user)"
+                >Save</a>
+              </div>
+              <a
+                class="theB delete"
+                @click="deleteUser(user.id)"
+              >Delete</a>
             </td>
           </tr>
         </tbody>
@@ -44,59 +64,62 @@
 
 <script>
 // @ is an alias to /src
-import { ipcRenderer } from 'electron'
-import Nav from './../components/Nav.vue';
+import { ipcRenderer } from 'electron';
+import Nav from '../components/Nav.vue';
+
 export default {
-  name: 'user',
+  name: 'User',
+  components: {
+    Nav,
+  },
   data() {
     return {
       editedUser: null,
-      tempArray: []
-    }
-  },
-  components: {
-      Nav
-  },
-  created() {
-    this.$store.dispatch('retrieveUsers')
+      tempArray: [],
+    };
   },
   computed: {
     allUsers() {
-      return this.$store.getters.allUsers
-    }
+      return this.$store.getters.allUsers;
+    },
   },
   watch: {
     anyBeacons() {
       if (this.tempArray.length <= 0) {
-        this.tempArray = this.$store.getters.allBeacons.splice(0)
-        this.$store.dispatch('retrieveBeacons')
-        console.log("reeee")
+        this.tempArray = this.$store.getters.allBeacons.splice(0);
+        this.$store.dispatch('retrieveBeacons');
+        console.log('reeee');
       }
-      console.log("test")
-      var tempArray = JSON.parse(JSON.stringify(this.tempArray))
-      var elderArray = JSON.parse(JSON.stringify(this.$store.getters.allBeacons))
-      console.log(tempArray, tempArray.length)
-      console.log(elderArray, elderArray.length)
+      console.log('test');
+      const tempArray = JSON.parse(JSON.stringify(this.tempArray));
+      const elderArray = JSON.parse(JSON.stringify(this.$store.getters.allBeacons));
+      console.log(tempArray, tempArray.length);
+      console.log(elderArray, elderArray.length);
       if (elderArray.length > tempArray.length) {
-        this.$store.dispatch('retrieveBeacons')
-        this.tempArray = this.$store.getters.allBeacons
+        this.$store.dispatch('retrieveBeacons');
+        this.tempArray = this.$store.getters.allBeacons;
       }
-    }
+    },
+  },
+  created() {
+    this.$store.dispatch('retrieveUsers');
   },
   methods: {
     openUserInterface() {
-      ipcRenderer.send('userInterface', 'open')
+      ipcRenderer.send('userInterface', 'open');
     },
     saveUser(user) {
-      this.$store.dispatch('updateUser', user)
-      this.editedUser = null
+      this.$store.dispatch('updateUser', user);
+      this.editedUser = null;
     },
     editUser(User) {
-      this.editedUser = User
+      this.editedUser = User;
     },
     deleteUser(id) {
-      this.$store.dispatch('deleteUser', id)
-    }
-  }
-}
+      if (confirm('Are you sure?')) {
+        this.$store.dispatch('deleteUser', id);
+      }
+    },
+  },
+};
 </script>
